+++ package.json
@@ -37,35 +37,35 @@
     "webpack": "^4.0.0"
   },
   "dependencies": {
-    "loader-utils": "^1.2.2",
-    "schema-utils": "^1.0.0"
+    "loader-utils": "^1.2.3",
+    "schema-utils": "^2.0.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.2.0",
-    "@babel/core": "^7.2.2",
-    "@babel/preset-env": "^7.2.0",
-    "@commitlint/cli": "^8.0.0",
-    "@commitlint/config-conventional": "^8.0.0",
-    "@webpack-contrib/defaults": "^5.0.0",
+    "@babel/cli": "^7.5.5",
+    "@babel/core": "^7.5.5",
+    "@babel/preset-env": "^7.5.5",
+    "@commitlint/cli": "^8.1.0",
+    "@commitlint/config-conventional": "^8.1.0",
+    "@webpack-contrib/defaults": "^5.0.2",
     "@webpack-contrib/eslint-config-webpack": "^3.0.0",
     "babel-jest": "^24.8.0",
     "commitlint-azure-pipelines-cli": "^1.0.2",
     "cross-env": "^5.2.0",
-    "del": "^4.1.1",
+    "del": "^5.0.0",
     "del-cli": "^2.0.0",
-    "eslint": "^5.16.0",
-    "eslint-config-prettier": "^4.3.0",
-    "eslint-plugin-import": "^2.17.3",
-    "husky": "^2.3.0",
+    "eslint": "^6.0.1",
+    "eslint-config-prettier": "^6.0.0",
+    "eslint-plugin-import": "^2.18.0",
+    "husky": "^3.0.0",
     "jest": "^24.8.0",
     "jest-junit": "^6.4.0",
-    "lint-staged": "^8.1.7",
+    "lint-staged": "^9.2.0",
     "memory-fs": "^0.4.1",
     "npm-run-all": "^4.1.5",
-    "prettier": "^1.17.1",
+    "prettier": "^1.18.2",
     "standard-version": "^6.0.1",
-    "url-loader": "^1.1.2",
-    "webpack": "^4.33.0"
+    "url-loader": "^2.0.1",
+    "webpack": "^4.36.1"
   },
   "keywords": [
     "webpack"
