+++ src/options.json
@@ -2,6 +2,7 @@
   "additionalProperties": true,
   "properties": {
     "name": {
+      "description": "The filename template for the target file(s) (https://github.com/webpack-contrib/file-loader#name).",
       "anyOf": [
         {
           "type": "string"
@@ -12,6 +13,7 @@
       ]
     },
     "outputPath": {
+      "description": "A filesystem path where the target file(s) will be placed (https://github.com/webpack-contrib/file-loader#outputpath).",
       "anyOf": [
         {
           "type": "string"
@@ -22,6 +24,7 @@
       ]
     },
     "publicPath": {
+      "description": "A custom public path for the target file(s) (https://github.com/webpack-contrib/file-loader#publicpath).",
       "anyOf": [
         {
           "type": "string"
@@ -32,12 +35,15 @@
       ]
     },
     "context": {
+      "description": "A custom file context (https://github.com/webpack-contrib/file-loader#context).",
       "type": "string"
     },
     "emitFile": {
+      "description": "Enables/Disables emit files (https://github.com/webpack-contrib/file-loader#emitfile).",
       "type": "boolean"
     },
     "regExp": {
+      "description": "A Regular Expression to one or many parts of the target file path. The capture groups can be reused in the name property using [N] placeholder (https://github.com/webpack-contrib/file-loader#regexp).",
       "anyOf": [
         {
           "type": "string"
@@ -47,14 +53,6 @@
         }
       ]
     }
-  },
-  "errorMessages": {
-    "name": "should be {String} or {Function} (https://github.com/webpack-contrib/file-loader#name)",
-    "outputPath": "should be {String} or {Function} (https://github.com/webpack-contrib/file-loader#outputpath)",
-    "publicPath": "should be {String} or {Function} (https://github.com/webpack-contrib/file-loader#publicpath)",
-    "context": "should be {String} (https://github.com/webpack-contrib/file-loader#context)",
-    "emitFile": "should be {Boolean} (https://github.com/webpack-contrib/file-loader#emitfile)",
-    "regExp": "should be {String} or {RegExp} (https://github.com/webpack-contrib/file-loader#regexp)"
   },
   "type": "object"
 }
